print temp on lcd

+++ inc/dma.h
@@ -1,9 +1,9 @@
+#include "MKL05Z4.h"
+
 #ifndef dma_h
 #define dma_h
 
-#include "MKL05Z4.h"
-
-extern uint16_t DMAvalue[8];
+extern uint32_t DMAvalue[8];
 
 void DMA_init();
